complete ex4.23

+++ backend/controllers/blog.js
@@ -1,8 +1,10 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
   const blogPosts = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   return response.json(blogPosts)
 })
 
@@ -15,15 +17,21 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-
+blogRouter.post('/', [middleware.tokenExtractor, middleware.userExtractor], async (request, response) => {
   if (!request.body.title && !request.body.url) {
     return response.status(400).json({ error: 'title and url must be provided'})
-  } else {
-    const savedBlogPost = await blog.save()
-    return response.status(201).json(savedBlogPost)
   }
+
+  const blog = new Blog({
+    ...request.body,
+    user: request.user._id.toString()
+  })
+
+  const savedBlogPost = await blog.save()
+  request.user.blogs = request.user.blogs.concat(savedBlogPost._id)
+  await request.user.save()
+
+  response.status(201).json(savedBlogPost)
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
@@ -40,9 +48,19 @@ blogRouter.put('/:id', async (request, response, next) => {
   }
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id',[middleware.tokenExtractor, middleware.userExtractor], async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === request.user._id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    request.user.blogs = request.user.blogs.pop(request.params.id)
+    await request.user.save()
+    response.status(204).end()
+  } else {
+    response.status(403).json({
+      error: `user ${user.username.toString()} does not have sufficient rights to delete post with ID ${request.params.id}`
+    })
+  }
 })
 
 module.exports = blogRouter
